+++ test/test_aozora2html.rb
@@ -77,7 +77,7 @@ class Aozora2HtmlTest < Test::Unit::TestCase
       begin
         assert_equal :kanji, parser.char_type(Embed_Gaiji_tag.new(nil,"foo","1-2-3","name"))
         assert_equal :kanji, parser.char_type(UnEmbed_Gaiji_tag.new(nil,"foo"))
-        assert_equal :hankaku, parser.char_type(Accent_tag.new(nil,123,"abc"))
+        assert_equal :hankaku, parser.char_type(Aozora2Html::Tag::Accent.new(nil,123,"abc"))
         assert_equal :else, parser.char_type(Okurigana_tag.new(nil,"abc"))
         assert_equal :else, parser.char_type(Inline_keigakomi_tag.new(nil,"abc"))
         assert_equal :katakana, parser.char_type(Dakuten_katakana_tag.new(nil,1,"abc"))
